backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sysatlas-backend"
version = "0.1.0"
description = "SysAtlas backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "cryptography>=44.0.0,<46.0.0",
  "fastapi>=0.115,<1.0",
  "pydantic-settings>=2.6,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "psycopg[binary]>=3.2,<4.0",
  "uvicorn[standard]>=0.32,<1.0"
]

[tool.setuptools.packages.find]
include = ["app*"]
